# Improve performance of automatic break regeneration